pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = "Tool for checking difference in json/yaml files"
authors = ["Pejnikov"]
repository = "https://github.com/Pejnikov/python-project-lvl2"
classifiers = [
    "License :: Free For Educational Use",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.7"
]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
mypy = "^0.910"
types-PyYAML = "^5.4.10"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
